[BUGFIX] displayCond IN operator doesn't work like documented
The FormEngine refactoring changed the representation
of field values, so that select boxes are now arrays.
The condition evaluation cannot handle those values as
prior to the refactoring and following the documentation.

This patch changes the behavior to follow strictly the
documentation. The = operator now compares only the first
element in case of an array. The IN operator compares two
potential lists for any match.

Fixes: #70421
Releases: master
Change-Id: I17e610215dede4c1c06c27d4d25ec1672bfccf83