[BUGFIX] Copy mode uses localize now but omits TransOrigPointer
While translating elements with the localization modal we have to
make sure that the "Copy" mode does more than just a plain copy
followed by an update of the sys_language_uid. Otherwise related
elements will just be copied along but don't get a new language.

So this patch replaces the copy action with a localize action
and just disables the usage of TransOrigPointer fields, so child
elements get the correct language but are still not connected to
a language parent.

Resolves: #75400
Releases: master, 7.6

Change-Id: If57af3c94c67824ab8bbbc62a8c6bea283e96fca